requirements.txt
numpy>=1.26.0
scikit-learn>=1.4.0
chromadb>=0.5.5
sentence-transformers>=3.0.1
numpy>=1.26
tqdm>=4.66
requests>=2.32
pydantic>=2.7
python-dotenv>=1.0.1

